refactor(plugins): more plugin config overhaul
