+++ lua/plugins/disabled.lua
@@ -5,6 +5,4 @@
 return {
   { 'shellRaining/hlchunk.nvim', enabled = true },
   { 'Bekaboo/deadcolumn.nvim', enabled = true },
-  { 'm4xshen/smartcolumn.nvim', enabled = false },
-  { 'goolord/alpha-nvim', enabled = false },
 }
