--- lua/plugins/utils/neo-tree.lua
@@ -1,75 +1,0 @@
-local util = require('core.util')
-
-return {
-  'nvim-neo-tree/neo-tree.nvim',
-  branch = 'v3.x',
-  dependencies = {
-    'nvim-lua/plenary.nvim',
-    'nvim-tree/nvim-web-devicons', -- not strictly required, but recommended
-    'MunifTanjim/nui.nvim',
-  },
-  keys = {
-    {
-      '<leader>fe',
-      function()
-        require('neo-tree.command').execute({ toggle = true, dir = util.get_root() })
-      end,
-      desc = 'Explorer (root dir)',
-    },
-    {
-      '<leader>fE',
-      function()
-        require('neo-tree.command').execute({ toggle = true, dir = vim.loop.cwd() })
-      end,
-      desc = 'Explorer (cwd)',
-    },
-    { '<leader>e', '<leader>fe', desc = 'Explorer (root dir)', remap = true },
-    { '<leader>E', '<leader>fE', desc = 'Explorer (cwd)', remap = true },
-  },
-  deactivate = function()
-    vim.cmd([[Neotree close]])
-  end,
-  init = function()
-    if vim.fn.argc() == 1 then
-      local stat = vim.loop.fs_stat(vim.fn.argv(0))
-      if stat and stat.type == 'directory' then
-        require('neo-tree')
-      end
-    end
-  end,
-  opts = {
-    sources = { 'filesystem', 'buffers', 'git_status', 'document_symbols' },
-    open_files_do_not_replace_types = { 'terminal', 'Trouble', 'qf', 'Outline' },
-    filesystem = {
-      bind_to_cwd = false,
-      follow_current_file = { enabled = true },
-      use_libuv_file_watcher = true,
-    },
-    window = {
-      mappings = {
-        ['<space>'] = 'none',
-      },
-    },
-    default_component_configs = {
-      indent = {
-        with_expanders = true, -- if nil and file nesting is enabled, will enable expanders
-        expander_collapsed = '',
-        expander_expanded = '',
-        expander_highlight = 'NeoTreeIndentMarker',
-      },
-    },
-  },
-  config = function(_, opts)
-    local function on_move(data)
-      util.on_rename(data.source, data.destination)
-    end
-
-    local events = require("neo-tree.events")
-    opts.event_handlers = opts.event_handlers or {}
-    vim.list_extend(opts.event_handlers, {
-      { event = events.FILE_MOVED, handler = on_move },
-      { event = events.FILE_RENAMED, handler = on_move },
-    })
-    require("neo-tree").setup(opts)
-  end,
-}
